Google Analytics setup.
Why?  I don't like tracking but rn the site is hosted by Netlify, which makes you pay for analytics.  Another option is to self-host or move to cloudflare pages, but that seems more work and the domain and dns is held by Netlify already.  Configured Analytics to only do basic page load tracking.